requirements.txt
opencv-python
mediapipe
python-dotenv
psycopg2
bcrypt